+++ index.html
@@ -28,5 +28,5 @@
   </div>
 </body>
 
-<script src="script.js"></script>
+<script type="module" src="./main.js"></script>
 </html>
